backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
botocore
pydantic
python-multipart
cors